ticket entity majorfeatures

+++ src/main/java/com/ticketing/ticketing_system/repositories/TicketRepository.java
@@ -3,5 +3,5 @@ package com.ticketing.ticketing_system.repositories;
 import com.ticketing.ticketing_system.entities.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<Ticket, Long> {
+public interface TicketRepository extends JpaRepository<Ticket, Integer> {
 }
